+++ packages/cdktf/lib/backends/index.ts
@@ -1,15 +1,15 @@
-export * from './backend_s3';
-export * from './backend_local';
-export * from './backend_remote';
-export * from './backend_artifactory';
-export * from './backend_azurerm';
-export * from './backend_consul';
-export * from './backend_cos';
-export * from './backend_etcd';
-export * from './backend_etcdv3';
-export * from './backend_gcs';
-export * from './backend_http';
-export * from './backend_manta';
-export * from './backend_oss';
-export * from './backend_pg';
-export * from './backend_swift';
+export * from './s3-backend';
+export * from './local-backend';
+export * from './remote-backend';
+export * from './artifactory-backend';
+export * from './azurerm-backend';
+export * from './consul-backend';
+export * from './cos-backend';
+export * from './etcd-backend';
+export * from './etcdv3-backend';
+export * from './gcs-backend';
+export * from './http-backend';
+export * from './manta-backend';
+export * from './oss-backend';
+export * from './pg-backend';
+export * from './swift-backend';
